% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct_data.R
\name{correct_blank}
\alias{correct_blank}
\title{Remove blank from each signal}
\usage{
correct_blank(data, blank_name = "Blank B", by_Group = FALSE)
}
\arguments{
\item{data}{A data frame or a tibble with at least a column 'Time', 'Signal',
'Content' and 'Value'.}

\item{blank_name}{A string corresponding to the name of the blank treatment in
'Content' column.}

\item{by_Group}{Correct the blank by group by using the 'Group' column if \code{TRUE}.}
}
\value{
A tibble with with same column as in data plus a column Value_BC (for
Blank Corrected) and column Blank
}
\description{
\code{correct_blank} Remove the background signal from the raw signal
}
\details{
\itemize{
\item 'Value' must be raw and not blank corrected.
\item If applying crosstalk correction for luminescence, it must be done before this step
}
}
\examples{
data <- Ecoli_T7_24well_Lux_2Fluo_OD
correct_blank(data, "Blank B", by_Group = TRUE)

# with crosstalk correction
crosstalk_lux_ref <- ct_calibration_24wells_4titude_black_vision
data_lux_corrected <- correct_crosstalk_lux(data, crosstalk_lux_ref, "LUX")
correct_blank(data_lux_corrected, "Blank B", by_Group = TRUE)
}
