% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/dataset_ct_calibration_24wells_4titude_black_vision.R
\docType{data}
\name{ct_calibration_24wells_4titude_black_vision}
\alias{ct_calibration_24wells_4titude_black_vision}
\title{Crosstalk calibration data for a black vision 4titude 24 wells (4ti-0241)}
\format{
\subsection{\code{ct_calibration_24wells_4titude_black_vision}}{

A list of two elements :
\describe{
\item{ct_matrix}{crosstalk matrix of 24x24 with coefficient of crosstalk
between all wells}
\item{blank}{mean blank value for this model and this medium}
}
}
}
\source{
https://cle.ird.fr/experiments.php?mode=view&id=1071
}
\usage{
ct_calibration_24wells_4titude_black_vision
}
\description{
A list with the crosstalk matrix for the model and the blank value in DM500
with a gain of 3000
}
\keyword{datasets}
