% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_Ecoli_T7_24well_Lux_2Fluo_OD.R
\docType{data}
\name{Ecoli_T7_24well_Lux_2Fluo_OD}
\alias{Ecoli_T7_24well_Lux_2Fluo_OD}
\title{Time kill curve E.Coli 10Cilux2 vs T7 4titude 24 wells (4ti-0241)}
\format{
\subsection{\code{Ecoli_T7_24well_Lux_2Fluo_OD}}{

A tibble of X columns :
\describe{
\item{X}{...}
\item{Y}{...}
}
}
}
\source{
https://cle.ird.fr//experiments.php?mode=view&id=164
}
\usage{
Ecoli_T7_24well_Lux_2Fluo_OD
}
\description{
A tibble with the raw values for each signals from the clariostar for each well
at each time.
}
\keyword{datasets}
