% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{import_excel}
\alias{import_excel}
\title{Tidy excel raw data}
\usage{
import_excel(path, name_repair = NULL)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{name_repair}{A list of the following format \code{list('raw_name_signal1' = 'new_name_signal1', 'raw_name_signal2' = 'new_name_signal2', ...)}. With the
'raw_name_signal' corresponding to the signal name in \emph{BMG MARS}. Default is
\code{NULL} and keep raw names from \emph{BMG MARS}.}
}
\value{
A tibble with columns : Well, Content, (Group,) Signal, Time, Value.
Signal contains either raw names from \emph{BMG MARS} or new names from \code{name_repair}.
}
\description{
\code{import_excel} import raw data exported from \emph{BMG MARS} into a tibble long
format with optionally signal names repaired.
}
\details{
In \emph{BMG MARS} you must choose these formats for your excel export :
\itemize{
\item hour format.
\item 'wide' format, meaning that time must go horizontally not vertically.
\item without 'metadata' before the table.
}
}
\examples{

# With repaired names
signal_names <- list("No filter 1" = "LUX", "480-20 2" = "CFP", "600-40 3" = "SYTOX", "4" = "OD")
import_excel(datasets, name_repair = signal_names)
}
