% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/microspectr-package.R
\docType{package}
\name{microspectr-package}
\alias{microspectr}
\alias{microspectr-package}
\title{microspectr: Process raw data from clariostar spectrophotometer}
\description{
A toolbox to 'tidy' raw signal data from the BMG clariostar spectrophotometer. You provide the data, tell 'microspectr' what signal and how it should analyse it and it take care of the details.
}
\author{
\strong{Maintainer}: Yoann Anciaux \email{anciaux.yoann@gmail.com} (\href{https://orcid.org/0000-0002-2845-4034}{ORCID})

}
\keyword{internal}
