% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_data.R
\name{microspectr_example}
\alias{microspectr_example}
\title{Get path to microspectr example}
\usage{
microspectr_example(path = NULL)
}
\arguments{
\item{path}{Name of file. If \code{NULL}, the example files will be listed.}
}
\description{
microspectr comes bundled with some example files in its \code{inst/extdata}
directory. This function make them easy to access.
}
\examples{
microspectr_example()
microspectr_example("Ecoli_T7_24well_Lux_2Fluo_OD.xlsx")
}
