% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{rescale_data}
\alias{rescale_data}
\title{Rescale the signal value by a reference}
\usage{
rescale_data(
  data,
  value_name = "Value_BC",
  rescale_from_well = NULL,
  rescale_from_control = NULL,
  control_name = "Positive control P",
  by_Group = FALSE
)
}
\arguments{
\item{data}{A data frame or a tibble with at least a column 'Time', 'Signal',
'Content' and 'Value'.}

\item{value_name}{A string corresponding to the name of the column containing
the value to be rescaled. Default is the output of \link{correct_blank} 'Value_BC'}

\item{rescale_from_well}{A named vector c('name_Signal1' = TimeForReferenceSignal1,
'name_Signal2' = TimeForReferenceSignal2, ...). TimeForReferenceSignalX is a time
existing in the column 'Time' at which the value of SignalX will be used as a
reference for rescaling. Default \code{NULL} leads to no rescaling by well.}

\item{rescale_from_control}{A named vector c('name_Signal1' = TimeForReferenceSignal1inControls,
'name_Signal2' = TimeForReferenceSignal2inControls, ...). TimeForReferenceSignalinControlsX
is a time existing in the column 'Time' at which the mean value of SignalX in
the controls will be used as a reference for rescaling. Default \code{NULL} leads
to no rescaling by controls.}

\item{control_name}{A string corresponding to the name of the controls in 'Content'.
Ony used if \code{rescale_from_control} is not \code{NULL}. Default "Positive control P".}

\item{by_Group}{A boolean. If \code{TRUE} the rescaling from the control is done by
'Group'. (Controls averaged by group)}
}
\value{
A tibble with with same column as in data plus a column Value_Rescaled
}
\description{
\code{rescale_data} Divide each 'Signal' by a avlue at a given time of the ame Signal,
either from a shared reference = the controls or a reference specific to each well.
}
\details{
\itemize{
\item If both \code{rescale_from_well} and \code{rescale_from_control} are \code{NULL} returns \code{NA}
in 'Value_Rescaled'.
\item Signals mentioned neither in \code{rescale_from_well} nor \code{rescale_from_control}
are \code{NA} in 'Value_Rescaled'
}
}
\examples{
data <- Ecoli_T7_24well_Lux_2Fluo_OD
rescale_from_well <- c("LUX" = 0, "OD" = 0, "SYTOX" = 5)
rescale_from_control <- c("CFP" = 10)
rescale_data(data, "Value", rescale_from_well, rescale_from_control,
             control_name = "Positive control P",  by_Group = TRUE)
}
