% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{tidy_data}
\alias{tidy_data}
\title{Import, correct and rescale data from \emph{BMG Clariostar}}
\usage{
tidy_data(
  path,
  name_repair = NULL,
  crosstalk_lux_ref = NULL,
  lux_name = NULL,
  blank_name = NULL,
  rescale_from_well = NULL,
  rescale_from_control = NULL,
  control_name = "Positive control P",
  by_Group = F
)
}
\arguments{
\item{path}{Path to the xls/xlsx file.}

\item{name_repair}{A list of the following format \code{list('raw_name_signal1' = 'new_name_signal1', 'raw_name_signal2' = 'new_name_signal2', ...)}. With the
'raw_name_signal' corresponding to the signal name in \emph{BMG MARS}. Default is
\code{NULL} and keep raw names from \emph{BMG MARS}.}

\item{crosstalk_lux_ref}{A list with two elements \code{list(ct_matrix = 'matrix of crosstalk', blank = 'value of the blank')}.}

\item{lux_name}{String with the name of the luminescence signal in the 'Signal'
column when providing a data with multiple signals. Default is NULL and will
consider that only the luminescent 'Value' are provided.}

\item{blank_name}{A string corresponding to the name of the blank treatment in
'Content' column.}

\item{rescale_from_well}{A named vector c('name_Signal1' = TimeForReferenceSignal1,
'name_Signal2' = TimeForReferenceSignal2, ...). TimeForReferenceSignalX is a time
existing in the column 'Time' at which the value of SignalX will be used as a
reference for rescaling. Default \code{NULL} leads to no rescaling by well.}

\item{rescale_from_control}{A named vector c('name_Signal1' = TimeForReferenceSignal1inControls,
'name_Signal2' = TimeForReferenceSignal2inControls, ...). TimeForReferenceSignalinControlsX
is a time existing in the column 'Time' at which the mean value of SignalX in
the controls will be used as a reference for rescaling. Default \code{NULL} leads
to no rescaling by controls.}

\item{control_name}{A string corresponding to the name of the controls in 'Content'.
Ony used if \code{rescale_from_control} is not \code{NULL}. Default "Positive control P".}

\item{by_Group}{Correct the blank by group by using the 'Group' column if \code{TRUE}.}
}
\value{
A tibble with columns as in the initial excel plus 'Signal', 'Time',
'Value', 'Blank', 'Value_BC', 'Value_Rescaled'.
}
\description{
\code{tidy_data} Is a wrapper for all the functions in 'microspectr' package. It
creates a streamline workflow to process a raw excel export froæ \emph{BMG MARS} to
data whcih can be directly fitted.
}
\details{
Particularities of the default values :
\itemize{
\item Default \code{crosstalk_lux_ref} is \code{NULL} which skips crosstalk correction.
\item Default \code{blank_name}is \code{NULL} which skips blank correction.
\item if both \code{rescale_from_well} and \code{rescale_from_control} are \code{NULL} it skips
the rescaling
}
}
\examples{
path <- microspectr_example("Ecoli_T7_24well_Lux_2Fluo_OD.xlsx")
name_repair <- list("No filter 1" = "LUX", "480-20 2" = "CFP", "600-40 3" = "SYTOX", "4" = "OD")
crosstalk_lux_ref <- ct_calibration_24wells_4titude_black_vision
rescale_from_well <- c("LUX" = 0, "OD" = 0, "SYTOX" = 5)
rescale_from_control <- c("CFP" = 10)
tidy_data(path = path, name_repair = name_repair,
          lux_name = "LUX", blank_name = "Blank B",
          rescale_from_well = rescale_from_well, rescale_from_control = rescale_from_control, control_name = "Positive control P",
          by_Group = TRUE)
}
